+++ packages/connectivity/src/scp-cf/destination/destination-service.ts
@@ -1,7 +1,8 @@
 import {
   createLogger,
   ErrorWithCause,
-  propertyExists
+  propertyExists,
+  removeTrailingSlashes
 } from '@sap-cloud-sdk/util';
 import CircuitBreaker from 'opossum';
 import axios, { AxiosRequestConfig, AxiosResponse, AxiosError } from 'axios';
@@ -17,7 +18,11 @@ import {
   DestinationJson,
   parseDestination
 } from './destination';
-import { Destination, DestinationType } from './destination-service-types';
+import {
+  Destination,
+  DestinationCertificate,
+  DestinationType
+} from './destination-service-types';
 import { destinationServiceCache } from './destination-service-cache';
 import { DestinationFetchOptions } from './destination-accessor-types';
 
@@ -37,7 +42,7 @@ type DestinationServiceOptions = ResilienceOptions &
   Pick<DestinationFetchOptions, 'destinationName'>;
 
 let circuitBreaker: DestinationCircuitBreaker<
-  DestinationJson | DestinationConfiguration
+  DestinationCertificate | DestinationConfiguration | DestinationJson
 >;
 
 /**
@@ -88,9 +93,8 @@ async function fetchDestinations(
   type: DestinationType,
   options?: DestinationsServiceOptions
 ): Promise<Destination[]> {
-  const targetUri = `${destinationServiceUri.replace(
-    /\/$/,
-    ''
+  const targetUri = `${removeTrailingSlashes(
+    destinationServiceUri
   )}/destination-configuration/v1/${type}Destinations`;
 
   if (options?.useCache) {
@@ -109,7 +113,7 @@ async function fetchDestinations(
 
   const headers = wrapJwtInHeader(jwt).headers;
 
-  return callDestinationService(targetUri, headers, options)
+  return callDestinationEndpoint(targetUri, headers, options)
     .then(response => {
       const destinations: Destination[] = response.data.map(d =>
         parseDestination(d)
@@ -165,14 +169,55 @@ export async function fetchDestination(
   );
 }
 
+/**
+ * Fetches a certificate from the subaccount and destination instance for a given a name.
+ * Subaccount is tried first.
+ * @param destinationServiceUri - The URI of the destination service
+ * @param token - The access token for destination service.
+ * @param certificateName - Name of the Certificate to be fetched
+ * @returns A Promise resolving to the destination
+ * @internal
+ */
+export async function fetchCertificate(
+  destinationServiceUri: string,
+  token: string,
+  certificateName: string
+): Promise<DestinationCertificate | undefined> {
+  const filetype = certificateName.split('.')[1];
+  if (filetype.toLowerCase() !== 'pem') {
+    logger.warn(
+      `The provided truststore ${certificateName} is not in 'pem' format which is currently the only supported format. Trustore is ignored.`
+    );
+    return;
+  }
+  const accountUri = `${removeTrailingSlashes(
+    destinationServiceUri
+  )}/destination-configuration/v1/subaccountCertificates/${certificateName}`;
+  const instanceUri = `${removeTrailingSlashes(
+    destinationServiceUri
+  )}/destination-configuration/v1/instanceCertificates/${certificateName}`;
+  const header = wrapJwtInHeader(token).headers;
+
+  try {
+    const response = await callCertificateEndpoint(accountUri, header).catch(
+      () => callCertificateEndpoint(instanceUri, header)
+    );
+    return response.data;
+  } catch (err) {
+    logger.warn(
+      `Failed to fetch truststore certificate ${certificateName} - Continuing without certificate. This may cause failing requests`,
+      err
+    );
+  }
+}
+
 async function fetchDestinationByTokens(
   destinationServiceUri: string,
   tokens: AuthAndExchangeTokens,
   options: DestinationServiceOptions
 ): Promise<Destination> {
-  const targetUri = `${destinationServiceUri.replace(
-    /\/$/,
-    ''
+  const targetUri = `${removeTrailingSlashes(
+    destinationServiceUri
   )}/destination-configuration/v1/destinations/${options.destinationName}`;
 
   let authHeader = wrapJwtInHeader(tokens.authHeaderJwt).headers;
@@ -184,7 +229,7 @@ async function fetchDestinationByTokens(
     ? { ...authHeader, 'X-tenant': tokens.exchangeTenant }
     : authHeader;
 
-  return callDestinationService(targetUri, authHeader, options)
+  return callDestinationEndpoint(targetUri, authHeader, options)
     .then(response => {
       const destination: Destination = parseDestination(response.data);
       return destination;
@@ -209,11 +254,45 @@ function errorMessageFromResponse(
     : '';
 }
 
-function callDestinationService(
+async function callCertificateEndpoint(
+  uri: string,
+  headers: Record<string, any>,
+  options?: ResilienceOptions
+): Promise<AxiosResponse<DestinationCertificate>> {
+  if (!uri.includes('Certificates')) {
+    throw new Error(
+      `callCertificateEndpoint was called with illegal arrgument: ${uri}. URL must be certificate endpoint of destination service.`
+    );
+  }
+  return callDestinationService(uri, headers, options) as Promise<
+    AxiosResponse<DestinationCertificate>
+  >;
+}
+
+async function callDestinationEndpoint(
   uri: string,
   headers: Record<string, any>,
   options?: ResilienceOptions
 ): Promise<AxiosResponse<DestinationJson | DestinationConfiguration>> {
+  if (!uri.match(/[instance|subaccount]Destinations|v1\/destinations/)) {
+    throw new Error(
+      `callDestinationEndpoint was called with illegal arrgument: ${uri}. URL must be destination(s) endpoint of destination service.`
+    );
+  }
+  return callDestinationService(uri, headers, options) as Promise<
+    AxiosResponse<DestinationConfiguration | DestinationJson>
+  >;
+}
+
+async function callDestinationService(
+  uri: string,
+  headers: Record<string, any>,
+  options?: ResilienceOptions
+): Promise<
+  AxiosResponse<
+    DestinationCertificate | DestinationConfiguration | DestinationJson
+  >
+> {
   const { enableCircuitBreaker, timeout } = {
     ...defaultResilienceBTPServices,
     ...options
@@ -227,19 +306,22 @@ function callDestinationService(
   };
 
   if (enableCircuitBreaker) {
-    return getCircuitBreaker().fire(config);
+    return getCircuitBreaker<
+      DestinationCertificate | DestinationJson | DestinationJson
+    >().fire(config);
   }
 
   return axios.request(config);
 }
 
-function getCircuitBreaker(): DestinationCircuitBreaker<
-  DestinationJson | DestinationConfiguration
+function getCircuitBreaker<T>(): DestinationCircuitBreaker<
+  DestinationCertificate | DestinationConfiguration | DestinationJson
 > {
   const request: (
     config: AxiosRequestConfig
-  ) => Promise<AxiosResponse<DestinationJson | DestinationConfiguration>> =
-    axios.request;
+  ) => Promise<
+    AxiosResponse<DestinationCertificate | DestinationJson | DestinationJson>
+  > = axios.request;
   if (!circuitBreaker) {
     circuitBreaker = new CircuitBreaker(request, circuitBreakerDefaultOptions);
   }