+++ CHANGELOG.md
@@ -67,7 +67,7 @@ Blog: TBD<br>
 
 ## New Functionality
 
-- [http-client] Introduce the `parameterEncoder` option to the request config of the `http-client` to allow custom parameter encoding.  
+- [http-client] Introduce the `parameterEncoder` option to the request config of the `http-client` to allow custom parameter encoding.
 - [http-client] Remove `method` from `defaultDisallowedKeys` to not filter out custom http method when using `filterCustomRequestConfig`.
 - [odata-common] Support adding custom http method in `addCustomRequestConfiguration` to overwrite the default http method.
 - [openapi] Support adding custom http method in `addCustomRequestConfiguration` to overwrite the default http method.